[EC-550] rename user admin view to org user admin view
