+++ apps/web/src/app/organizations/core/views/index.ts
@@ -2,4 +2,4 @@ export * from "./collection-access-selection.view";
 export * from "./collection-admin.view";
 export * from "./group.view";
 export * from "./organization-user.view";
-export * from "./user-admin-view";
+export * from "./organization-user-admin-view";
